CHANGE : move version.go to internal version pkg .
